requirements.txt
opencv-python
numpy
python-dotenv